move core to AixWeather
