+++ AixWeather/core_data_format_2_output_file/unconverted_to_x.py
@@ -6,8 +6,8 @@ import json
 import pickle
 import pandas as pd
 
-from core.core_data_format_2_output_file import utils_2output
-from core.imports.utils_import import MetaData
+from AixWeather.core_data_format_2_output_file import utils_2output
+from AixWeather.imports.utils_import import MetaData
 
 
 def to_pickle(core_df: pd.DataFrame, meta: MetaData):
