+++ AixWeather/imports/custom_file.py
@@ -3,7 +3,7 @@ change this file to your custom requirements (only import and perform as
 little preprocessing as possible)
 """
 
-from core.imports.utils_import import MetaData
+from AixWeather.imports.utils_import import MetaData
 
 
 def load_custom_meta_data():
