+++ AixWeather/transformation_to_core_data/DWD.py
@@ -1,14 +1,9 @@
 from datetime import datetime, timedelta
 import pandas as pd
 
-from core.imports.utils_import import MetaData
-from core.transformation_functions import (
-    auxiliary,
-    time_observation_transformations,
-    unit_conversions,
-    variable_transformations,
-    pass_through_handling,
-)
+from AixWeather.imports.utils_import import MetaData
+from AixWeather.transformation_functions import auxiliary, time_observation_transformations, variable_transformations, \
+    pass_through_handling, unit_conversions
 
 """
 format_DWD_historical information:
