+++ AixWeather/transformation_to_core_data/ERC.py
@@ -4,14 +4,9 @@ This module includes functions to transform ERC data to core data format.
 
 import pandas as pd
 
-from core.imports.utils_import import MetaData
-from core.transformation_functions import (
-    time_observation_transformations,
-    unit_conversions,
-    variable_transformations,
-    auxiliary,
-    pass_through_handling,
-)
+from AixWeather.imports.utils_import import MetaData
+from AixWeather.transformation_functions import auxiliary, time_observation_transformations, variable_transformations, \
+    pass_through_handling, unit_conversions
 
 """
 format_ERC information
